Fixed GCC Wconversion warnings 11/23/2022 | 22w47b1

+++ Games/Sandbox/src/ParticleSystem2D/ParticleSystem2DLayer.h
@@ -53,8 +53,8 @@ public:
 			const TRAP::Graphics::OrthographicCameraBounds bounds = m_cameraController.GetBounds();
 			const TRAP::Math::Vec3 pos = m_cameraController.GetCamera().GetPosition();
 
-			const float x = (mousePosition.x / resolution.x) * bounds.GetWidth() - bounds.GetWidth() * 0.5f;
-			const float y = bounds.GetHeight() * 0.5f - (mousePosition.y / resolution.y) * bounds.GetHeight();
+			const float x = (mousePosition.x / static_cast<float>(resolution.x)) * bounds.GetWidth() - bounds.GetWidth() * 0.5f;
+			const float y = bounds.GetHeight() * 0.5f - (mousePosition.y / static_cast<float>(resolution.y)) * bounds.GetHeight();
 			m_particle.Position = { x + pos.x, y + pos.y };
 			for (uint32_t i = 0; i < 5; i++)
 				m_particleSystem.Emit(m_particle);
