+++ Games/Tests/src/RendererAPI/RendererAPITests.cpp
@@ -236,7 +236,7 @@ bool RendererAPITests::OnKeyPress(TRAP::Events::KeyPressEvent& e)
 	}
 	if(e.GetKey() == TRAP::Input::Key::F4)
 	{
-		m_pushConstantOrUBO = (m_pushConstantOrUBO + 1) % 3;
+		m_pushConstantOrUBO = static_cast<uint8_t>((m_pushConstantOrUBO + 1) % 3);
 		TP_TRACE("[RendererAPITests] Push Constant / Uniform Buffer: ", m_pushConstantOrUBO ? "On" : "Off");
 	}
 	if(e.GetKey() == TRAP::Input::Key::V)
