+++ TRAP/src/Layers/ImGui/ImGuiLayer.cpp
@@ -427,7 +427,9 @@ bool ImGui::ImageButton(TRAP::Ref<TRAP::Graphics::Texture> image, const ImVec2& 
 		TRAP::Utils::HashCombine(imgViewHash, imgView);
 		const ImGuiID imgViewID = static_cast<uint32_t>(imgViewHash);
 
-		return ImGui::ImageButtonEx(imgViewID, texID, size, uv0, uv1, ImVec2(frame_padding, frame_padding), bg_col, tint_col);
+		return ImGui::ImageButtonEx(imgViewID, texID, size, uv0, uv1,
+		                            ImVec2(static_cast<float>(frame_padding),
+									static_cast<float>(frame_padding)), bg_col, tint_col);
 	}
 
 	return false;
